Use javadoc Op declaration instead
