--- scijava-ops-tutorial/src/main/resources/META-INF/services/org.scijava.ops.spi.OpCollection
@@ -1,11 +1,0 @@
-# -- DO NOT EDIT -- This file is autogenerated
-# Instead, modify and re-run <basedir>/bin/generate-meta-inf.sh from the top-level
-org.scijava.ops.tutorial.OpAdaptation
-org.scijava.ops.tutorial.OpConversion
-org.scijava.ops.tutorial.OpDependencies
-org.scijava.ops.tutorial.OpParallelization
-org.scijava.ops.tutorial.OpPriorities
-org.scijava.ops.tutorial.OpReduction
-org.scijava.ops.tutorial.ReportingProgress
-org.scijava.ops.tutorial.UsingNils
-org.scijava.ops.tutorial.WritingOpCollections
